T8154: make catch_broken_pipe decorator to common vyos.utils.io function
This adds the capability to make it re-usable in other code paths.
