+++ python/vyos/utils/io.py
@@ -111,3 +111,14 @@ def select_entry(l: list, list_msg: str = '', prompt_msg: str = '',
     select = ask_input(prompt_msg, default=default_entry, numeric_only=True,
                        valid_responses=valid_entry)
     return next(filter(lambda x: x[0] == select, en))[1]
+
+def catch_broken_pipe(func):
+    import os
+    import sys
+    def wrapped(*args, **kwargs):
+        try:
+            func(*args, **kwargs)
+        except (BrokenPipeError, KeyboardInterrupt):
+            # Flush output to /dev/null and bail out.
+            os.dup2(os.open(os.devnull, os.O_WRONLY), sys.stdout.fileno()) # pylint: disable = no-member
+    return wrapped
